[arcane,accelerator] Renomme 'RunQueueInternal.h' en 'KernelLauncher.h'.
Supprime aussi 'RunQueueInternal.cc' qui est inutile.
