+++ arcane/src/arcane/accelerator/srcs.cmake
@@ -15,6 +15,7 @@ set( ARCANE_SOURCES
   GenericScanner.h
   GenericSorter.h
   IReduceMemoryImpl.h
+  KernelLauncher.h
   MaterialVariableViews.h
   MaterialVariableViews.cc
   MemoryCopier.cc
@@ -31,7 +32,6 @@ set( ARCANE_SOURCES
   Runner.h
   RunQueue.h
   RunQueueImpl.h
-  RunQueueInternal.cc
   RunQueueInternal.h
   Filter.h
   Filterer.cc
